requirements.txt
django==2.2
django-bootstrap4
django-heroku
gunicorn
dj-database-url
dj-static
whitenoise
psycopg2-binary==2.8
markdown
pillow
djangorestframework
django-filter